+++ src/NEFDashboard/components/addUser/index.jsx
@@ -10,9 +10,9 @@ import { getRole } from "../../../utils";
 import { useTranslation } from "react-i18next";
 import "./addUser.css";
 
-const AddUser = ({ open, handleSubmit, onClose }) => {
+const AddUser = ({ open, handleSubmit, onClose, selectedUser = {} }) => {
   const { user } = useAuth();
-  const adminApi = new adminDashboardApi(user.roles && user.roles[0]);
+  const adminApi = new adminDashboardApi(user.roles && getRole(user));
   const { t } = useTranslation();
   const [roles, setRoles] = useState([
     { value: "Vendor", label: t("dashboard.roles.Vendor") },
@@ -34,6 +34,29 @@ const AddUser = ({ open, handleSubmit, onClose }) => {
     service_fee: "1",
   });
 
+  function capitalizeFirstLetter(str) {
+    return str.charAt(0).toUpperCase() + str.slice(1);
+  }
+
+  const capitalizeAllWords = (str) => {
+    const words = str.split(" ");
+    const capitalizedWords = words.map(capitalizeFirstLetter);
+    return capitalizedWords.join(" ");
+  };
+
+  useEffect(() => {
+    if (Object.keys(selectedUser)?.length > 0)
+      setUserDetail({
+        firstName: selectedUser?.firstName,
+        lastName: selectedUser?.lastName,
+        email: selectedUser?.email,
+        service_fee: "1",
+        company: selectedUser?.business,
+        role: capitalizeAllWords(selectedUser?.roles[0]),
+        // password: "",
+      });
+  }, [selectedUser]);
+
   const handleUserDetails = (value, key) => {
     setUserDetail({ ...userDetail, [key]: value });
   };
@@ -41,7 +64,9 @@ const AddUser = ({ open, handleSubmit, onClose }) => {
   const handleSubmitUser = async () => {
     if (
       Object.keys(userDetail)?.some(
-        (key) => userDetail[key] == "" && key != "company",
+        (key) =>
+          (userDetail[key] == "" && key != "company") ||
+          (Object.keys(selectedUser)?.length > 0 && key == "password"),
       )
     ) {
       setErrorMessage("All fields are required");
@@ -57,10 +82,25 @@ const AddUser = ({ open, handleSubmit, onClose }) => {
     delete payload.role;
     delete payload.service_fee;
 
-    const resp = await adminApi.createUser(payload);
+    const resp =
+      Object.keys(selectedUser)?.length > 0
+        ? await adminApi.updateUser(
+            payload?.firstName,
+            payload?.lastName,
+            payload?.email,
+            payload?.email,
+            userDetail?.role,
+          )
+        : await adminApi.createUser(payload);
     if (resp) {
       if (resp.ok) {
-        setInfoMessage(t("messages.success.addUser"));
+        setInfoMessage(
+          t(
+            Object.keys(selectedUser)?.length > 0
+              ? "Updated Successfully"
+              : "messages.success.addUser",
+          ),
+        );
         handleSubmit();
       } else if (resp.status === 409) {
         setErrorMessage(t("messages.error.userExist"));
@@ -116,7 +156,9 @@ const AddUser = ({ open, handleSubmit, onClose }) => {
         <Flex align="center" gap={4} className="add-modal-title-container">
           <img src={AddUserIcon} alt="add_user" />
           <div className="default-text-gray add-user_modal_title">
-            Add New User
+            {Object.keys(selectedUser)?.length > 0
+              ? t("dashboard.modal.edit_user")
+              : t("dashboard.modal.titleCreate")}
           </div>
         </Flex>
       }
@@ -178,6 +220,7 @@ const AddUser = ({ open, handleSubmit, onClose }) => {
               className="add-user-input-fields"
               value={userDetail?.password}
               onChange={(e) => handleUserDetails(e.target.value, "password")}
+              disabled={Object.keys(selectedUser)?.length > 0}
             />
           </Flex>
           <Flex vertical gap={6}>
@@ -282,7 +325,9 @@ const AddUser = ({ open, handleSubmit, onClose }) => {
           className="default-text add-user-submit cursor-pointer"
           onClick={handleSubmitUser}
         >
-          {t("dashboard.modal.titleCreate")}
+          {Object.keys(selectedUser)?.length > 0
+            ? t("update")
+            : t("dashboard.modal.titleCreate")}
         </div>
       </Flex>
     </Modal>